pyproject.toml
[project]
name = "chatbot-text2sql-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "accelerate>=0.25.0",
    "arize-phoenix==4.5.0",
    "beautifulsoup4>=4.14.2",
    "chainlit==2.9.0",
    "chromadb>=1.3.5",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langfuse==2.20.0",
    "langgraph==1.0.3",
    "markdownify>=1.2.2",
    "openai==2.8.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "pandas==2.3.3",
    "plotly==6.4.0",
    "python-dotenv==1.2.1",
    "requests>=2.32.5",
    "sentence-transformers==2.2.2",
    "tiktoken>=0.12.0",
    "torch>=2.6.0",
    "transformers>=4.43.2",
    "vllm>=0.8.5",
]
